us_foodscope/health/serving/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
scikit-learn==1.3.2
joblib==1.3.2
numpy==1.26.3
pydantic==2.5.3
python-multipart
pandas
faiss-cpu
sentence-transformers
groq
huggingface_hub
python-dotenv
